requirements.txt
alembic==1.12.1
asyncpg==0.29.0

# Development Tools
black==24.3.0
boto3==1.34.0
celery==5.3.4
# Web Framework
fastapi==0.109.2
flake8==6.1.0
greenlet==3.2.3

# HTTP Client
httpx==0.25.2
isort==5.12.0

# File Storage
minio==7.2.11
mypy==1.7.1

# Data Processing
numpy==1.24.4
pandas==2.0.3
passlib[bcrypt]==1.7.4
pre-commit==3.6.0

# System Monitoring
psutil==5.9.6
psycopg2-binary==2.9.9
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Environment & Configuration
python-dotenv==1.0.0

# Authentication
python-jose[cryptography]==3.3.0
python-multipart==0.0.18

# Redis
redis==5.0.1

# Database
sqlalchemy==2.0.23

# Type Stubs
types-redis==4.6.0.8
types-requests==2.31.0.10
uvicorn[standard]==0.24.0
